update: user login required

+++ base/views.py
@@ -2,9 +2,10 @@ from multiprocessing import context
 from re import I
 from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
-from django.contrib.auth import authenticate, login
+from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
@@ -31,6 +32,11 @@ def loginPage(request):
     coontext = {}
     return render(request, 'base/login_register.html', coontext)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
 
@@ -52,7 +58,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,10 +70,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed her!!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -77,9 +87,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed her!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
